backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=13.0,<15.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
google-genai>=0.1.0
google-generativeai>=0.8.0
openai>=1.23.6
# pydantic v2 fix - 2025-12-10
